--- a/Cargo.toml
+++ b/Cargo.toml
@@ -13,6 +13,7 @@ num-traits = "0.2"
 rayon = "1.5"
 rand = { version = "0.9.0", features = ["std_rng"] }
 itertools = "0.14.0"
+memory-stats = "1.0.0"
 
 [dev-dependencies]
 keccak-asm = { version = "0.1.4" }