--- a/src/poly/matrix.rs
+++ b/src/poly/matrix.rs
@@ -90,4 +90,6 @@ pub trait PolyMatrix:
         &self,
         new_modulus: &<<Self::P as Poly>::Params as PolyParams>::Modulus,
     ) -> Self;
+    /// Performs the operation S * (identity ⊗ other)
+    fn mul_tensor_identity(&self, other: &Self, identity_size: usize) -> Self;
 }