--- a/src/utils.rs
+++ b/src/utils.rs
@@ -1,11 +1,11 @@
-use num_bigint::BigUint;
-use num_traits::{One, Zero};
-
 use crate::poly::{
     dcrt::{DCRTPoly, DCRTPolyParams, DCRTPolyUniformSampler},
     sampler::DistType,
     Poly,
 };
+use memory_stats::memory_stats;
+use num_bigint::BigUint;
+use num_traits::{One, Zero};
 
 pub fn ceil_log2(q: &BigUint) -> usize {
     assert!(!q.is_zero(), "log2 is undefined for zero");
@@ -68,5 +68,13 @@ pub fn create_bit_poly(params: &DCRTPolyParams, bit: bool) -> DCRTPoly {
         DCRTPoly::const_one(params)
     } else {
         DCRTPoly::const_zero(params)
+    }
+}
+
+pub fn print_memory_usage(label: &str) {
+    if let Some(usage) = memory_stats() {
+        println!("{}: {} bytes", label, usage.physical_mem);
+    } else {
+        println!("Couldn't get memory stats!");
     }
 }