Core idea: I have a `S` matrix of size 2x12, `other` matrix of size 3x3 and identity `I` matrix of size 4x4. Goal is to perform `S * (I ⊗ other)`.

Current implementation: First perform `partial = (I ⊗ other)` and then `out = S * partial`.

Optimized implementation: Directly compute `out = S1*other | S2*other | S3*other | S4*other` where each `Si` is the slice of `S` of size 2 x 3

Results after running `/usr/bin/time -l cargo test test_io_just_mul_enc_and_bit --release  -- --nocapture`

Before
```
Peak memory consumptions 4801904640  bytes (4.47 GB)
Time to obfuscate: 132.466593875s
Time for evaluation: 129.248957959s
Total time: 261.715551834s
```

After
```
Peak memory consumptions 542539776  bytes (0.51 GB)
Time to obfuscate: 105.465769375s
Time for evaluation: 116.751426959s
Total time: 220.925412584s
```